LPD-86520 Extract Demandbase logic into its own class
